requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
ftfy==6.1.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
Pillow==9.5.0
regex==2023.5.5
requests==2.30.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.65.0
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.3
zipp==3.15.0
